tink: move tests to be integration tests
